requirements.txt
hydra-core==1.3.2
matplotlib==3.6.2
nltk==3.8.1
omegaconf==2.2.3
pydantic==1.10.2
scikit_learn==1.3.0
torchtext==0.15.2
tqdm==4.64.1
pandas==2.0.3
--index-url https://download.pytorch.org/whl/cu118
torch
torchaudio
torchvision
--index-url https://pypi.org/simple
pytorch-lightning==2.0.1.post0
torchmetrics==0.11.4
